Commit Jakob
I slightly  adjusted exercise 6, and added a note explaining the time frame. A printed the df with absolute spread values for exercise 10 in addition to the chart. I also took out the missing value interplotation for exercise 9, as our code now handles the missing values well beforehand and therefore I adjusted exercise 10 to make use of the merged dataset.